Fixed performance issue when scrolling down for the first time in a long document.
The Highlighter is calling OnHighlightStateChanged() once for each line in the document when the highlighting is initialized for the first time. For each of those calls, we were calling Redraw() to invalidate the following line.
Now we avoid these unnecessary Redraw() calls by ignoring all OnHighlightStateChanged() calls for all 'scanned' lines (all lines other than the one currently being colorized).