Alter line selection painting
This removes the "extra" selection visible at the start of the next line, referenced in issue  #199. The existing paint logic is inconsistent with Visual Studio (and just about all other editors, AFAIK). I have been using this code in my fork for over a year, and it appears to work fine. But note that I haven't tried to address corner cases. For example, if word wrap is turned on, and the last character selected is a space, and the line breaks on that space, the fix won't apply (a fairly extreme corner case). There may be other corner cases that aren't addressed, such as an image on the last line, etc.